Revert "FEAT: User와 Company 도메인과 엔티티를 생성한다."
