+++ build.gradle
@@ -59,17 +59,7 @@ dependencies {
 	//feign Client
 	implementation 'org.springframework.cloud:spring-cloud-starter-openfeign:4.1.5'
 
-	//SpringSecurity
-	implementation "org.springframework.boot:spring-boot-starter-security"
 
-	// Security
-	implementation 'org.springframework.boot:spring-boot-starter-security'
-	testImplementation 'org.springframework.security:spring-security-test'
-
-	//JWT
-	implementation 'io.jsonwebtoken:jjwt-api:0.12.5'
-	runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.12.5'
-	runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.12.5'
 
 }
