+++ src/main/java/hanieum/conik/ConikApplication.java
@@ -3,11 +3,9 @@ package hanieum.conik;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @EnableFeignClients
 @SpringBootApplication
-@EnableJpaAuditing
 public class ConikApplication {
 
 	public static void main(String[] args) {
