--- src/main/java/hanieum/conik/adapter/member/dto/MemberLoginRequest.java
@@ -1,13 +1,0 @@
-package hanieum.conik.adapter.member.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record MemberLoginRequest(
-        @NotBlank(message = "이메일은 필수 입력입니다.")
-        @Email(message = "유효한 이메일 형식이어야 합니다.")
-        String email,
-
-        @NotBlank(message = "비밀번호는 필수 입력입니다.")
-        String password
-) {}
