--- src/main/java/hanieum/conik/adapter/member/dto/MemberLoginResponse.java
@@ -1,7 +1,0 @@
-package hanieum.conik.adapter.member.dto;
-
-public record MemberLoginResponse(
-        String accessToken,
-        String refreshToken,
-        Long memberId
-) { }
