--- src/main/java/hanieum/conik/adapter/member/dto/MemberSignUpResponse.java
@@ -1,5 +1,0 @@
-package hanieum.conik.adapter.member.dto;
-
-public record MemberSignUpResponse(
-) {
-}
