--- src/main/java/hanieum/conik/application/member/MemberFinderService.java
@@ -1,25 +1,0 @@
-package hanieum.conik.application.member;
-
-import hanieum.conik.application.member.provided.MemberFinder;
-import hanieum.conik.application.member.required.MemberRepository;
-import hanieum.conik.domain.member.Member;
-import hanieum.conik.domain.member.exception.MemberErrorType;
-import hanieum.conik.domain.member.exception.MemberException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.validation.annotation.Validated;
-
-@Service
-@Transactional
-@Validated
-@RequiredArgsConstructor
-public class MemberFinderService implements MemberFinder {
-
-    private final MemberRepository memberRepository;
-
-    @Override
-    public Member find(Long memberId) {
-        return memberRepository.findById(memberId).orElseThrow(() -> new MemberException(MemberErrorType.MEMBER_NOT_FOUND));
-    }
-}
