+++ src/main/java/hanieum/conik/domain/user/exception/UserErrorType.java
@@ -1,0 +1,17 @@
+package hanieum.conik.domain.user.exception;
+
+import hanieum.conik.global.apiPayload.exception.ErrorType;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public enum UserErrorType implements ErrorType {
+    INVALID_AUTHORIZATION_CODE(HttpStatus.UNAUTHORIZED, "유효하지 않은 인증 코드입니다."),
+    ;
+
+    private final HttpStatus status;
+
+    private final String message;
+}
