--- src/main/java/hanieum/conik/global/adapter/security/CurrentUser.java
@@ -1,14 +1,0 @@
-package hanieum.conik.global.adapter.security;
-
-import io.swagger.v3.oas.annotations.Parameter;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-@Parameter(hidden = true)
-public @interface CurrentUser {
-}
